requirements.txt
tensorflow==2.15.0
flask==3.0.3
flask-cors==4.0.0
flask-limiter==3.5.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
gunicorn==21.2.0
